+++ GarmentsManagement/src/main/java/com/emranhss/GarmentsManagement/repository/DistrictRepository.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface DistrictRepository extends JpaRepository<District,Integer> {
+public interface DistrictRepository extends JpaRepository<District,Long> {
 
-    List<District> findByDivisionId(Integer divisionId);
+    List<District> findByDivisionId(Long divisionId);
 
     List<District> findByDivisionName(String divisionName);
 }
