+++ GarmentsManagement/src/main/java/com/emranhss/GarmentsManagement/service/DivisionService.java
@@ -12,10 +12,10 @@ import java.util.Optional;
 public interface DivisionService {
     Division save(Division d);
     List<Division> findAll();
-    Optional<Division> getById(Integer id);
-    void delete(Integer id);
+    Optional<Division> getById(Long id);
+    void delete(Long id);
 
-    List<DivisionDto> getDivisionsByCountryId(Integer countryId);
+    List<DivisionDto> getDivisionsByCountryId(Long countryId);
 
     List<DivisionDto> getDivisionsByCountryName(String countryName);
